Add option to hide footer
I'm using this class to create a front page with TOC to include as the first page in the merged document.

To do that, I first need to merge my documents to get the TOC object. 
Then I'm using the TOC object and another plugin to generate the front page. 
After that, I'm merging once more. 

But I can't have the footer generated both times, because it'll overlap. So I'm proposing a small change to allow the user to choose whether the footer should be generated.